Added a few type signatures, for documentation, and more importantly, to make the library compile with ghc-7.* ...
Explanation: ghc-7.* introduces a new form of "monomorphism
restriction". When type families and/or GADTs are used, then
local bindings are by default not generalized. Since the Core
module makes use of polymorphic local functions in a few places,
I added type signatures for them in order to make the typechecker
happy.

git-svn-id: https://svn.science.uu.nl/repos/project.STEC.uu-parsinglib/trunk@72 b22e4a66-0d8d-4a73-aa21-d6b4dd4dcce4